refactor: extract DocumentationLoader, remove hardcoded app regexes (#119)
- Extract DocumentationLoader from OutputComprehender.ts into its own
  file src/agents/comprehension/DocumentationLoader.ts with a clean
  public interface and no app-specific defaults.
- Remove hardcoded atg / azure-tenant-grapher CLI regex patterns from
  DocumentationLoader.extractFeatures(); replace with a configurable
  cliCommandPatterns?: RegExp[] field on ComprehensionAgentConfig.
- Update OutputComprehender.ts to re-export DocumentationLoader for
  backward compatibility.
- Update comprehension/index.ts barrel to export DocumentationLoader
  directly from its own module.
- Update ComprehensionAgent.ts to pass config.cliCommandPatterns to
  DocumentationLoader constructor.
- Update tests: CLI-pattern tests now supply patterns explicitly;
  add tests for the new cliCommandPatterns config field.
- 27 tests pass, tsc --noEmit clean.

Closes #119
